Feat: 로그인 확인 로직 추가 인터셉터로 처리하면 인터셉터 내부의 로직이 너무 복잡해져, 단순하게 컨트롤러에 일단 추가

+++ src/main/java/codesquad/codestagram/controller/ReplyController.java
@@ -3,6 +3,7 @@ package codesquad.codestagram.controller;
 import codesquad.codestagram.domain.Reply;
 import codesquad.codestagram.domain.User;
 import codesquad.codestagram.dto.ReplyResponseDto;
+import codesquad.codestagram.exception.NotLoggedInException;
 import codesquad.codestagram.service.ReplyService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +21,10 @@ public class ReplyController {
     @PostMapping("/articles/{index}/answers")
     public ResponseEntity<ReplyResponseDto> addAnswer(@PathVariable Long index, @RequestParam String text, HttpSession session) {
         User user = (User) session.getAttribute("loginUser");
+        if (user == null) {
+            throw new NotLoggedInException("로그인이 필요합니다.");
+        }
+
         Reply reply = replyService.addReplyToArticle(user, index, text);
         return ResponseEntity.ok(ReplyResponseDto.from(reply));
     }
@@ -27,6 +32,10 @@ public class ReplyController {
     @DeleteMapping("/articles/{articleId}/answers/{replyId}")
     public ResponseEntity<?> addAnswer(@PathVariable Long articleId, @PathVariable Long replyId, HttpSession session) {
         User user = (User) session.getAttribute("loginUser");
+        if (user == null) {
+            throw new NotLoggedInException("로그인이 필요합니다.");
+        }
+
         Reply reply = replyService.findReplyIfOwner(user,replyId);
         replyService.deleteReply(reply);
         return ResponseEntity.noContent().build();
